Allow creation of files through API with async job
The prior handling used the Item Builder to add a file, then got that
file out of the database and set the other data from the API request to
it: this works fine if the process of adding a file is totally
synchronous but has issues if async processing is used. In particular,
it's possible/likely for data (really, just the order, if given) to be
overwritten by the background process working on the file and its
thumbnails.

To avoid this problem, we can just pass in the metadata to the initial
builder call to add the file, and then the API doesn't need to look up
the File record again to add anything, so there's no timing/overwriting
issues. This implies some plumbing changes:

- allow sending the order to the builder call (and therefore the file
  ingests)
- allow the Upload ingest (used by the API) to take extra data like
  element texts and now "order"
- refactor the API record adapter to avoid duplication of code for
  getting the element texts from the API request